JWTAuthDRF/requirements.txt
Django
djangorestframework_simplejwt
djangorestframework